Use View::make() instead of calling a function that encapsulates this
By using View::make(), Confide users are able to alter their UserController to pass data to their views using ->with('page_title', 'This is passed to my view') when needed. This is useful for passing data to a user's view such as a page title or meta description for a master template. When the View::make() for the form is encapsulated in a function call, this is not possible. 