Update README.md
I have this replacement view for the login form:




    @extends('layouts.master')
    
    @section('content')
    {{{ Confide::makeLoginForm()->render() }}}
    @stop




It looks like this in the html source:




![layout-confide](https://f.cloud.github.com/assets/661038/321518/fdb0a366-99e2-11e2-9140-f240241ac82d.png)




It looks like it's related to [this merge](https://github.com/laravel/framework/issues/206) in Laravel 4. Suprisingly, it works as expected with two brackets instead:




    {{{ Confide::makeLoginForm()->render() }}}




Which seems to be just the opposite of what the comments say. So this just updates the docs to reflect what actually works (I updated Laravel just yesterday so I have the latest build).