Prevent span.help-inline around signup errors
In the login view it returns a string in ```Session::get('error')``` but in register it returns an array populated from validation. In Laravel 4 this wraps with ```<span class="help-inline"></span>```. This was making my ```div.alert``` look weird.

This pull removes any default error wrappers to be more consistent with how the Login form receives error text.