Update ConfideUser.php
Depend the transport of the mail regarding the config-setting. If wanted we can log it when it is set to false. But now the email is only send when you specifically want it, true or false in the config-file. Some scenarios require you not to send out the mail before something is paid for example.