Fixes minor bootstrap 3 issues with login form.
Fixes the floating checkbox over the label per the basic example provided in Bootstrap 3's documentation. 

The default bootstrap CSS applies a negative left margin to the checkbox, so the box should come before the label text. Also, the containing div should be classed as a "checkbox" not "form-control." Also added the 'role="form"' to the form declaration and moved the "forgot password" link to a paragraph classed "help-block" below the password entry for consistency with Bootstrap 3 documentation.